sysbench: add new recipe
The project used to be hosted on Sourceforge, then Launchpad, and it moved to
Github. The last stable release (0.4.12) cannot be found on Sourceforge anymore,
and is not available (as tarball) on github. So using the tarball from Launchpad
instead. We can move to github when the next release happens (which will be the
first one from github).

Tested with and without mysql, with and without aio.

Signed-off-by: Nicolas Dechesne <nicolas.dechesne@linaro.org>
Signed-off-by: Martin Jansa <Martin.Jansa@gmail.com>